# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AlphaSwarm Control Plane",
    "description": "Multi-tenant algorithmic trading SaaS API",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/strategies": {
      "post": {
        "tags": [
          "strategies"
        ],
        "summary": "Create Strategy",
        "operationId": "create_strategy_v1_strategies_post",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "strategies"
        ],
        "summary": "List Strategies",
        "operationId": "list_strategies_v1_strategies_get",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StrategyResponse"
                  },
                  "title": "Response List Strategies V1 Strategies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/strategies/{strategy_id}": {
      "get": {
        "tags": [
          "strategies"
        ],
        "summary": "Get Strategy",
        "operationId": "get_strategy_v1_strategies__strategy_id__get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/strategies/{strategy_id}/runs": {
      "post": {
        "tags": [
          "strategies"
        ],
        "summary": "Run Strategy",
        "operationId": "run_strategy_v1_strategies__strategy_id__runs_post",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "strategies"
        ],
        "summary": "Get Task Status",
        "operationId": "get_task_status_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "rejected"
        ],
        "title": "RunStatus"
      },
      "StrategyCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Name"
          },
          "prompt": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 10,
            "title": "Prompt"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symbols"
          },
          "timeframe": {
            "type": "string",
            "maxLength": 20,
            "title": "Timeframe",
            "default": "1Min"
          },
          "risk": {
            "$ref": "#/components/schemas/StrategyRiskConfig"
          }
        },
        "type": "object",
        "required": [
          "name",
          "prompt"
        ],
        "title": "StrategyCreateRequest"
      },
      "StrategyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "owner_user_id": {
            "type": "string",
            "title": "Owner User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symbols"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/StrategyStatus"
              }
            ],
            "default": "draft"
          },
          "generated_logic": {
            "type": "string",
            "title": "Generated Logic"
          },
          "risk": {
            "$ref": "#/components/schemas/StrategyRiskConfig"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "owner_user_id",
          "name",
          "prompt",
          "symbols",
          "timeframe",
          "generated_logic",
          "risk"
        ],
        "title": "StrategyResponse"
      },
      "StrategyRiskConfig": {
        "properties": {
          "max_order_notional": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Max Order Notional",
            "default": 1000.0
          },
          "max_daily_notional": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Max Daily Notional",
            "default": 5000.0
          },
          "allowed_symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Symbols"
          },
          "paper_trading_only": {
            "type": "boolean",
            "title": "Paper Trading Only",
            "default": true
          }
        },
        "type": "object",
        "title": "StrategyRiskConfig"
      },
      "StrategyRunRequest": {
        "properties": {
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          }
        },
        "type": "object",
        "title": "StrategyRunRequest"
      },
      "StrategyRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "strategy_id": {
            "type": "string",
            "title": "Strategy Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "strategy_id",
          "task_id",
          "status",
          "dry_run",
          "message"
        ],
        "title": "StrategyRunResponse"
      },
      "StrategyStatus": {
        "type": "string",
        "enum": [
          "draft",
          "active",
          "paused",
          "archived"
        ],
        "title": "StrategyStatus"
      },
      "TaskStatusResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "celery_status": {
            "type": "string",
            "title": "Celery Status"
          },
          "result": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "celery_status"
        ],
        "title": "TaskStatusResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}